Move Email and Mailer to the most appropriate location
Since the modules are used by the service notifier, move them to the
service_notifier folder
